........
  r28543 | kohsuke | 2010-03-12 16:07:55 -0800 (Fri, 12 Mar 2010) | 1 line
  
  fixed a regression caused by HUDSON-5780
........
  r28544 | kohsuke | 2010-03-12 16:09:25 -0800 (Fri, 12 Mar 2010) | 1 line
  
  fixed a regression caused by HUDSON-5780
........
  r28545 | kohsuke | 2010-03-12 16:24:14 -0800 (Fri, 12 Mar 2010) | 1 line
  
  fixed a regression caused by HUDSON-5780
........
  r28546 | kohsuke | 2010-03-12 16:52:36 -0800 (Fri, 12 Mar 2010) | 1 line
  
  [maven-release-plugin] prepare release hudson-1_350
........
  r28548 | kohsuke | 2010-03-12 16:52:58 -0800 (Fri, 12 Mar 2010) | 1 line
  
  [maven-release-plugin] prepare for next development iteration
........
  r28551 | kohsuke | 2010-03-12 17:36:02 -0800 (Fri, 12 Mar 2010) | 1 line
  
  updated changelog as a part of the release
........
  r28625 | abayer | 2010-03-15 09:54:14 -0700 (Mon, 15 Mar 2010) | 1 line
  
  Fixing regression that resulted in all but latest build artifacts getting deleted
........
  r28627 | kohsuke | 2010-03-15 10:09:52 -0700 (Mon, 15 Mar 2010) | 1 line
  
  touch up
........
  r28628 | kohsuke | 2010-03-15 10:10:23 -0700 (Mon, 15 Mar 2010) | 1 line
  
  another pointless array conversion
........
  r28629 | kohsuke | 2010-03-15 10:45:13 -0700 (Mon, 15 Mar 2010) | 1 line
  
  [maven-release-plugin] prepare release hudson-1_351
........
  r28631 | kohsuke | 2010-03-15 10:45:50 -0700 (Mon, 15 Mar 2010) | 1 line
  
  [maven-release-plugin] prepare for next development iteration
........
  r28635 | kohsuke | 2010-03-15 11:29:03 -0700 (Mon, 15 Mar 2010) | 1 line
  
  updated changelog as a part of the release
........


git-svn-id: https://hudson.dev.java.net/svn/hudson/trunk/hudson/main@28639 71c3de6d-444a-0410-be80-ed276b4c234a

+++ core/src/main/java/hudson/slaves/NodeProvisioner.java
@@ -221,7 +221,7 @@ public class NodeProvisioner {
         }
     }
 
-    private static final float MARGIN = (float)(Integer.getInteger(NodeProvisioner.class.getName()+".MARGIN",10)/100);
+    private static final float MARGIN = Integer.getInteger(NodeProvisioner.class.getName()+".MARGIN",10)/100f;
     private static final Logger LOGGER = Logger.getLogger(NodeProvisioner.class.getName());
 
     // TODO: picker should be selectable
